rename create button to create_poly button
