+++ label_coach/web_client/src/control/create_poly_button/presenter.jsx
